Pgadmin activated & Minimal updates in controller layer

+++ src/main/java/com/demo/app/api/rest/controllers/CustomerController.java
@@ -46,9 +46,8 @@ public class CustomerController {
 			@QueryParam("phone_num") String phoneNum,
 			@QueryParam("email") String email,
 			@QueryParam("is_active") Boolean isActive,
-			@QueryParam("discount_rate") Integer discountRate,
+			@QueryParam("discount_rate") Double discountRate,
 			@QueryParam("order_id") Integer orderId) {
-		// HashMap<String, Object> dataMap = customerService.getDataMap();
 		HashMap<String, Object> dataMap = new HashMap<String, Object>();
 		dataMap.put("name", name);
 		dataMap.put("surname", surname);
